added Difficulty enum, and small changes
