+++ question-service/src/main/resources/db/data/1.0/difficulty_enum.sql
@@ -1,1 +1,1 @@
-# CREATE TYPE public.Difficulty AS ENUM ('HARD', 'EASY', 'MEDIUM');
+CREATE TYPE public.Difficulty AS ENUM ('HARD', 'EASY', 'MEDIUM');
