+++ topic-service/src/main/java/com/information/center/topicservice/controller/TopicController.java
@@ -2,7 +2,7 @@ package com.information.center.topicservice.controller;
 
 import com.information.center.topicservice.model.request.TopicRequest;
 import com.information.center.topicservice.model.response.TopicResponse;
-import com.information.center.topicservice.service.TopicService;
+import com.information.center.topicservice.service.TopicServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -14,37 +14,43 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
 public class TopicController implements TopicEndpoint {
 
-    private final TopicService topicService;
+    private final TopicServiceImpl topicServiceImpl;
 
+    @Override
     @PostMapping
     public TopicResponse create(@RequestBody TopicRequest topicResponse) {
 
-        return topicService.create(topicResponse);
+        return topicServiceImpl.create(topicResponse);
     }
 
+    @Override
     @GetMapping("/internal/{externalId}")
     public String getNameById(@PathVariable("externalId") String externalId) {
-        return topicService.getNameById(externalId);
+        return topicServiceImpl.getNameById(externalId);
     }
 
+    @Override
     @PutMapping
     public void update(@RequestBody TopicResponse topicResponse) {
-        topicService.update(topicResponse);
+        topicServiceImpl.update(topicResponse);
     }
 
+    @Override
     @GetMapping("/{externalId}")
     public TopicResponse findByExternalId(@PathVariable("externalId") String externalId) {
-        return topicService.findByExternalId(externalId);
+        return topicServiceImpl.findByExternalId(externalId);
     }
 
+    @Override
     @GetMapping
     public Page<TopicResponse> findAll(Pageable pageable) {
-        return topicService.findAll(pageable);
+        return topicServiceImpl.findAll(pageable);
     }
 
+    @Override
     @DeleteMapping("/{externalId}")
     public void delete(@PathVariable("externalId") String externalId) {
 
-        topicService.delete(externalId);
+        topicServiceImpl.delete(externalId);
     }
 }
