+++ video-service/src/main/java/com/information/center/videoservice/controller/VideoController.java
@@ -19,27 +19,33 @@ public class VideoController implements VideoEndpoint {
 
     private final VideoService videoService;
 
+    @Override
     public VideoResponse findById(@PathVariable("externalId") String externalId) {
         return videoService.findByExternalId(externalId);
     }
 
+    @Override
     public VideoResponse create(VideoRequest videoRequest) throws IOException {
         return videoService.create(videoRequest);
     }
 
+    @Override
     public void update(VideoDto videoDto) throws IOException {
         videoService.update(videoDto);
     }
 
+    @Override
     public void delete(@PathVariable("externalId") String externalId) {
         videoService.delete(externalId);
     }
 
+    @Override
     public ResponseEntity<UrlResource> getFullVideo(@PathVariable("externalId") String externalId) throws MalformedURLException {
         return videoService.getFullVideo(externalId);
 
     }
 
+    @Override
     public List<VideoResponse> findAllByTopicId(@PathVariable("topicId") String topicId) {
         return videoService.findAllByTopicId(topicId);
     }
